# re-use text escaper